first test -- creating reusable ImageUpload component
