+++ src/components/Otsu.vue
@@ -15,6 +15,7 @@
   <div class="Otsu">
     <h1>{{ msg }}</h1>
     <v-container class="text-xs-left">
+
       <v-layout row wrap>
         <v-flex xs12>
           <h3>Description</h3>
@@ -32,6 +33,7 @@
             DO NOT ZIP the folder containing images; select all images and ZIP them directly.</p>
         </v-flex>
       </v-layout>
+
       <v-alert
         v-model="loginRequired"
         type="error"
@@ -39,6 +41,7 @@
       >
         {{loginRequiredMsg}}
       </v-alert>
+
       <v-alert
         v-model="errorAlert"
         type="error"
@@ -46,6 +49,7 @@
       >
         {{errorAlertMsg}}
       </v-alert>
+
       <v-dialog v-model="successDlg" persistent max-width="500px">
         <v-card>
           <v-card-title>
@@ -60,41 +64,19 @@
           </v-card-actions>
         </v-card>
       </v-dialog>
-      <v-flex xs12 class="text-xs-center text-sm-center text-md-center text-lg-center">
-        <p class="text-xs-left">Select File
-          <v-btn class="text-xs-left" small color="primary" @click='pickFile'>Browse</v-btn>
-          <input
-            type="file"
-            style="display: none"
-            accept=".jpg, .png, .tif, .mat, .zip"
-            ref="myUpload"
-            @change="onFilePicked"
-          >
-        </p>
-        <v-list v-model="fileName" subheader: true v-if="fileUploaded">
-          <v-list-tile
-            v-for="file in filesDisplay"
-            :key="file.fileName"
-          >
-            <v-list-tile-avatar>
-              <v-icon color="primary">check_circle_outline</v-icon>
-            </v-list-tile-avatar>
-            <v-list-tile-content>
-              <v-list-tile-title v-text="file.fileName"></v-list-tile-title>
-            </v-list-tile-content>
-            <v-btn v-on:click="openImageEditor()" color="primary">Edit image</v-btn>
-            <EditImage v-model='imageEditorOpen' v-bind:img='file' v-on:setCroppedImage="setCroppedImage"></EditImage>
-          </v-list-tile>
-        </v-list>
-      </v-flex>
+
+      <ImageUpload v-on:setFiles="setFiles"></ImageUpload>
+
       <v-flex class="text-xs-center">
-      <v-btn id="binarize-button" v-on:click="submit()" color="primary">Binarize</v-btn>
+        <v-btn id="binarize-button" v-on:click="submit()" color="primary">Binarize</v-btn>
       </v-flex>
+
       <h4>References</h4>
       <v-flex xs12>
         <p> N. Otsu, A threshold selection method from gray-level histograms, IEEE transactions on systems, man, and
           cybernetics, vol. 9, no. 1, pp. 62-66, 1979.</p>
       </v-flex>
+
     </v-container>
   </div>
 </template>
@@ -103,32 +85,28 @@
 import {} from 'vuex'
 import {JobMgr} from '@/modules/JobMgr.js'
 import {Auth} from '@/modules/Auth.js'
-import EditImage from './EditImage.vue'
+import ImageUpload from './ImageUpload.vue'
 
 export default {
   name: 'Otsu',
   components: {
-    EditImage
+    ImageUpload
   },
   data: () => {
     return ({
       title: 'Input Upload',
       msg: 'Binarization - Otsu Method',
       dialog: false,
-      fileName: '',
-      // file_type: [],
-      files: [],
-      filesDisplay: [],
       errorAlert: false,
       errorAlertMsg: '',
       loginRequired: false,
       loginRequiredMsg: '',
-      fileUploaded: false,
       successDlg: false,
       jobId: '',
-      imageEditorOpen: false
+      files: [],
     })
   },
+
   beforeMount: function () {
     let vm = this
     vm.auth = new Auth()
@@ -137,21 +115,11 @@ export default {
       vm.loginRequiredMsg = 'Login is required.'
     }
   },
-  methods: {
-    openImageEditor: function () {
-      this.imageEditorOpen = !this.imageEditorOpen // toggle the image editor modal being open and closed
-    },
 
-    setCroppedImage: function (...args) {    
-      for (let i = 0; i < this.files.length; i++){
-        if (this.files[i].fileName == args[1].fileName){
-          this.files[i].fileUrl = args[0];
-          this.filesDisplay[i].fileUrl = args[0];
-          this.filesDisplay[i].fileName = "cropped_" + this.filesDisplay[i].fileName; // to ensure that the list of images reloads since they are binded to filenames.
-          console.log('image succesfully cropped.')
-          return;
-        }
-      }
+  methods: {
+
+    setFiles: function (files) {
+      this.files = files;
     },
 
     setLoading: function () {
@@ -162,46 +130,12 @@ export default {
       this.$store.commit('notLoading')
     },
 
-    pickFile () {
-      this.$refs.myUpload.click()
-    },
-
-    resetFiles: function () {
-      this.files = []
-      this.filesDisplay = []
-      this.fileUploaded = false
-    },
-
-    onFilePicked (e) {
-      this.resetFiles()
-      const files = e.target.files
-      for (let i = 0; i < files.length; i++) {
-        let file = {}
-        let f = files[i]
-        if (f !== undefined) {
-          file.fileName = f.name
-          if (file.fileName.lastIndexOf('.') <= 0) {
-            return
-          }
-          console.log(file.fileName)
-          const fr = new FileReader()
-          fr.readAsDataURL(f)
-          fr.addEventListener('load', () => {
-            file.fileUrl = fr.result
-            this.files.push(file)
-            this.filesDisplay.push(file)
-            this.fileUploaded = true
-          })
-        } else {
-          console.log('File Undefined')
-        }
-      }
-    },
     successDlgClicked: function () {
       let vm = this
       console.log('Success dlg button clicked')
       vm.$router.go(-2) // go back to mcr homepage page
     },
+
     submit: function () {
       let vm = this
       vm.files.forEach(function (v) {
